meta: move messageHandler
